+++ tests/testthat/test_space_filling.R
@@ -12,14 +12,14 @@ test_that('max entropy designs', {
   )
   expect_equal(nrow(grid_1), 11L)
   expect_true(all(grid_1$mixture > 0 & grid_1$mixture < 1))
-  expect_true(all(grid_1$cost > -10 & grid_1$cost < -1))
+  expect_true(all(grid_1$cost > -10 & grid_1$cost < 5))
 
   grid_2 <- grid_max_entropy(
     cost(), mixture(),
     size = 11,
     original = TRUE
   )
-  expect_true(all(grid_2$cost > 2^-10 & grid_2$cost < 2^-1))
+  expect_true(all(grid_2$cost > 2^-10 & grid_2$cost < 2^5))
 
   grid_3 <- grid_max_entropy(
     cost(),
@@ -59,14 +59,14 @@ test_that('latin square designs', {
   )
   expect_equal(nrow(grid_1), 11L)
   expect_true(all(grid_1$mixture > 0 & grid_1$mixture < 1))
-  expect_true(all(grid_1$cost > -10 & grid_1$cost < -1))
+  expect_true(all(grid_1$cost > -10 & grid_1$cost < 5))
 
   grid_2 <- grid_latin_hypercube(
     cost(), mixture(),
     size = 11,
     original = TRUE
   )
-  expect_true(all(grid_2$cost > 2^-10 & grid_2$cost < 2^-1))
+  expect_true(all(grid_2$cost > 2^-10 & grid_2$cost < 2^5))
 
   grid_3 <- grid_latin_hypercube(
     cost(),